+++ ms-quartos/src/main/java/br/com/fiap/hackathon/quartos/controller/PredioController.java
@@ -1,19 +1,25 @@
 package br.com.fiap.hackathon.quartos.controller;
 
+import br.com.fiap.hackathon.quartos.Exception.GenericException;
 import br.com.fiap.hackathon.quartos.dtos.PredioDto;
+import br.com.fiap.hackathon.quartos.entity.Localidade;
 import br.com.fiap.hackathon.quartos.entity.Predio;
+import br.com.fiap.hackathon.quartos.entity.Quarto;
 import br.com.fiap.hackathon.quartos.mappers.PredioMapper;
+import br.com.fiap.hackathon.quartos.mappers.QuartoMapper;
 import br.com.fiap.hackathon.quartos.service.LocalidadeService;
 import br.com.fiap.hackathon.quartos.service.PredioService;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import br.com.fiap.hackathon.quartos.service.QuartoService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import jakarta.validation.Valid;
 import org.springframework.http.ResponseEntity;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
@@ -21,14 +27,18 @@ import org.springframework.web.bind.annotation.*;
 public class PredioController {
   private final PredioService predioService;
   private final PredioMapper predioMapper;
-
   private final LocalidadeService localidadeService;
+  private final QuartoService quartoService;
+  private final QuartoMapper quartoMapper;
 
   public PredioController(
-      PredioService predioService, PredioMapper predioMapper, LocalidadeService localidadeService) {
+      PredioService predioService, PredioMapper predioMapper, LocalidadeService localidadeService,
+      QuartoService quartoService,  QuartoMapper quartoMapper) {
     this.predioService = predioService;
     this.predioMapper = predioMapper;
     this.localidadeService = localidadeService;
+    this.quartoService =  quartoService;
+    this.quartoMapper = quartoMapper;
   }
 
   @GetMapping
@@ -58,6 +68,10 @@ public class PredioController {
         @ApiResponse(description = "Não encontrado", responseCode = "404")
       })
   public ResponseEntity<PredioDto> getPredioById(@PathVariable String id) {
+    if(!predioService.existsById(id)){
+      throw new GenericException("Não exite um predio cadastrado com esse ID: " + id + "! Não é possível continuar!");
+    }
+
     return ResponseEntity.ok(predioMapper.toDto(predioService.getPredioById(id)));
   }
 
@@ -72,12 +86,28 @@ public class PredioController {
       })
   public ResponseEntity<PredioDto> createPredio(@RequestBody @Valid PredioDto predioDto) {
     if (!localidadeService.existsById(predioDto.getLocalidadeId())) {
-
-      return ResponseEntity.badRequest().body(null);
+      throw new GenericException("Localidade com ID " + predioDto.getLocalidadeId() + " não encontrado! Cadastre uma localidade com este ID antes de cadastrar um novo Predio!");
     }
 
+    if(predioService.existsById(predioDto.getId())){
+      throw new GenericException("Já exite um predio cadastrado com esse ID: " + predioDto.getId() + "! Não é possível continuar!");
+    }
+
+    if(!predioDto.getQuartos().stream().allMatch(quarto -> quarto.getPredioId().equals(predioDto.getId()))){
+      throw new GenericException("Id do predio informado diferente do Id do predio informado em algum dos quartos!");
+    }
+
+    predioDto.getQuartos().stream().forEach(quarto -> {
+      if(quartoService.existsById(quarto.getId())){
+        throw new GenericException("Já existe um quarto cadastrado com esse ID: " + quarto.getId() + "! não é possível continuar!");
+      }
+    });
+  //Busca a localidade, faz o insert do predio no banco, adiciona o predio na localida e da update atualizar a localidade.
+    Localidade localidade = localidadeService.getLocalidadeById(predioDto.getLocalidadeId());
     Predio predio = predioMapper.toEntity(predioDto);
-    predio = predioService.createPredio(predio);
+    predioService.createPredio(predio);
+    localidade.getPredios().add(predio);
+    localidadeService.updateLocalidade(predioDto.getLocalidadeId(), localidade);
 
     return ResponseEntity.ok(predioMapper.toDto(predio));
   }
@@ -94,6 +124,10 @@ public class PredioController {
       })
   public ResponseEntity<PredioDto> updatePredio(
       @PathVariable String id, @RequestBody PredioDto predioDto) {
+    if(!predioService.existsById(id)){
+      throw new GenericException("Não exite um predio cadastrado com esse ID: " + id + "! Não é possível continuar!");
+    }
+
     Predio predio = predioMapper.toEntity(predioDto);
     predio = predioService.updatePredio(id, predio);
     return ResponseEntity.ok(predioMapper.toDto(predio));
@@ -107,6 +141,9 @@ public class PredioController {
         @ApiResponse(description = "Não encontrado", responseCode = "404")
       })
   public ResponseEntity<Void> deletePredio(@PathVariable String id) {
+    if(!predioService.existsById(id)){
+      throw new GenericException("Não exite um predio cadastrado com esse ID: " + id + "! Não é possível continuar!");
+    }
     predioService.deletePredio(id);
     return ResponseEntity.ok().build();
   }
